moved types to models package
