+++ README.md
@@ -1,2 +1,3 @@
 # piquel-cli
+
 CLI tool to manage my machine. Basically just a wrapper for shell scripts and commands
